[REF] test_config_settings
