+++ odoo/addons/test_orm/models/__init__.py
@@ -7,7 +7,6 @@ from . import (
     test_autovacuum,
     test_check_access,
     test_company_checks,
-    test_config,
     test_config_settings,
     test_convert,
     test_convert_env,
